refactor(gateway): Q11.f — extract rest_v1/tasks/activity.rs (GAR-655)
mod.rs: 2071 → 1919 lines (−152). New activity.rs (173 LOC) holds
ActivityRow/ActivityResponse/ListActivityQuery/ListActivityResponse structs
and list_task_activity handler. Pure structural refactor — zero behavior change.
All items re-exported via `pub use activity::{...}` in mod.rs.
